requirements/base.txt
Django==2.0.6
django-js-asset==1.1.0
django-mptt==0.9.0
djangorestframework==3.8.2
djangorestframework-gis==0.13
factory-boy==2.11.1
Faker==0.8.16
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.4
text-unidecode==1.2
